feat : mypage api

+++ dangmoog/lib/screens/mypage/my_page.dart
@@ -16,6 +16,7 @@ import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 
 class MyPage extends StatefulWidget {
   final String email;
+
   const MyPage({super.key, required this.email, required String nickname});
 
   @override
@@ -31,14 +32,46 @@ Future<double?> tillGetSource(Stream<double> source) async {
   return null; // No positive value found
 }
 
+Future<String?> fetchProfileImageUrl() async {
+  try {
+    final Response response = await ApiService().autoLogin(); // API 호출
+
+    if (response.statusCode == 200) {
+      final Map<String, dynamic> data = response.data;
+      final String? profileUrl = data["profile_url"]; // "profile_url" 값을 가져옴
+
+      if (profileUrl != null) {
+        return profileUrl;
+      } else {
+        // "profile_url"이 null인 경우 처리
+        return null;
+      }
+    } else {
+      // API 응답에 문제가 있는 경우 오류 처리
+      throw Exception('API 응답 오류: ${response.statusCode}');
+    }
+  } catch (e) {
+    print('이미지 URL 가져오기 오류: $e');
+    return null; // 에러 발생 시 null 반환
+  }
+}
+
 class _MyPageState extends State<MyPage> {
   final storage = const FlutterSecureStorage();
+  //late Future<String?> profileImageUrl; // 프로필 이미지 URL을 저장할 변수
+
+  @override
+  void initState() {
+    super.initState();
+    // profileImageUrl = fetchProfileImageUrl(); // 프로필 이미지 URL 가져오기
+  }
 
   @override
   Widget build(BuildContext context) {
     String userEmail = Provider.of<UserProvider>(context).inputEmail;
     String userNickname = Provider.of<UserProvider>(context).nickname;
     File? userImage = Provider.of<UserProvider>(context).userImage;
+
     return FutureBuilder<double?>(
       future: tillGetSource(Stream<double>.periodic(
           const Duration(milliseconds: 100),
@@ -61,18 +94,39 @@ class _MyPageState extends State<MyPage> {
                 Row(children: [
                   SizedBox(width: screenSize.width * 0.042),
                   ClipOval(
-                    //borderRadius: BorderRadius.circular(50),
-                    child: userImage != null // UserProvider에서 이미지를 가져옴
-                        ? Image.file(
-                            userImage,
-                            width: screenSize.width * 0.14,
-                            height: screenSize.width * 0.14,
-                            fit: BoxFit.cover,
+                    child: fetchProfileImageUrl() != null
+                        ? FutureBuilder<String?>(
+                            future:
+                                fetchProfileImageUrl(), // fetchImage 함수 호출하여 profileUrl을 가져옴
+                            builder: (context, snapshot) {
+                              if (snapshot.connectionState ==
+                                  ConnectionState.done) {
+                                final profileUrl = snapshot.data;
+                                if (profileUrl != null) {
+                                  return Image.network(
+                                    profileUrl,
+                                    width: screenSize.width * 0.14,
+                                    height: screenSize.width * 0.14,
+                                    fit: BoxFit.cover,
+                                  );
+                                } else {
+                                  return Image.asset(
+                                    'assets/images/basic_profile.png',
+                                    width: screenSize.width * 0.14,
+                                    height: screenSize.width * 0.14,
+                                    fit: BoxFit.cover,
+                                  );
+                                }
+                              } else if (snapshot.hasError) {
+                                return Text('Error: ${snapshot.error}');
+                              } else {
+                                return CircularProgressIndicator(); // 데이터 로딩 중 표시
+                              }
+                            },
                           )
                         : Image.asset(
-                            'assets/images/sample.png', // 기본 이미지
-                            width: screenSize.width * 0.14,
-                            height: screenSize.width * 0.14,
+                            'assets/images/basic_profile.png',
+                            fit: BoxFit.cover,
                           ),
                   ),
                   SizedBox(width: screenSize.width * 0.042),
@@ -267,8 +321,16 @@ class _MyPageState extends State<MyPage> {
                 MypageText(
                     text: '로그아웃',
                     icon: Icons.logout_outlined,
-                    onPressed: () async {
-                      try {
+                    onPressed: () {
+                      LogoutPopup(screenSize, context);
+                    }),
+                MypageText(
+                  text: '탈퇴하기',
+                  icon: Icons.delete_outline_outlined,
+                  onPressed: () async {
+                    try {
+                      Response response = await ApiService().deleteAccount();
+                      if (response.statusCode == 204) {
                         await storage.delete(key: 'accessToken');
                         await storage.delete(key: 'userId');
 
@@ -278,17 +340,61 @@ class _MyPageState extends State<MyPage> {
                               builder: (context) => const WelcomePage()),
                           (route) => false,
                         );
-                      } catch (e) {
-                        print("로그아웃에 실패했습니다.");
                       }
-                    }),
-                MypageText(
-                  text: '탈퇴하기',
-                  icon: Icons.delete_outline_outlined,
+                    } catch (e) {
+                      print(e);
+                    }
+                  },
+                ),
+              ],
+            ),
+          );
+        }
+        return Container();
+      },
+    );
+  }
+
+  Future LogoutPopup(
+    Size screenSize,
+    BuildContext context,
+  ) {
+    return showDialog(
+      context: context,
+      builder: (BuildContext context) {
+        return AlertDialog(
+          backgroundColor: Color(0xFFFFFFFF),
+          shape: RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(14.0),
+          ),
+          title: Column(
+            children: [
+              const Text(
+                '로그아웃',
+                textAlign: TextAlign.center,
+                style: TextStyle(
+                  fontSize: 16,
+                  fontWeight: FontWeight.w600,
+                  color: Color(0xFF302E2E),
+                ),
+              ),
+            ],
+          ),
+          content: const Text(
+            '정말 로그아웃 하시겠어요?',
+            textAlign: TextAlign.center,
+            style: TextStyle(
+              fontSize: 14,
+              fontWeight: FontWeight.w400,
+              color: Color(0xFF302E2E),
+            ),
+          ),
+          actions: [
+            Column(
+              children: [
+                ElevatedButton(
                   onPressed: () async {
-                    Response response = await ApiService().deleteAccount();
-                    print(response);
-                    if (response.statusCode == 200) {
+                    try {
                       await storage.delete(key: 'accessToken');
                       await storage.delete(key: 'userId');
 
@@ -298,14 +404,62 @@ class _MyPageState extends State<MyPage> {
                             builder: (context) => const WelcomePage()),
                         (route) => false,
                       );
+                    } catch (e) {
+                      print("로그아웃에 실패했습니다.");
                     }
                   },
+                  style: ElevatedButton.styleFrom(
+                    backgroundColor: const Color(0xFFE20529),
+                    shape: RoundedRectangleBorder(
+                      borderRadius: BorderRadius.circular(6),
+                    ),
+                    padding:
+                        const EdgeInsets.symmetric(horizontal: 12, vertical: 4),
+                    minimumSize: Size(
+                      screenSize.width * 0.67,
+                      screenSize.height * 0.044,
+                    ),
+                  ),
+                  child: const Text(
+                    '로그아웃',
+                    style: TextStyle(
+                      fontSize: 13,
+                      fontWeight: FontWeight.w400,
+                      color: Color(0xFFFFFFFF),
+                    ),
+                  ),
+                ),
+                ElevatedButton(
+                  onPressed: () {
+                    // 팝업 창을 닫을 때 수행할 작업을 여기에 추가하세요.
+                    Navigator.of(context).pop();
+                  },
+                  style: ElevatedButton.styleFrom(
+                    backgroundColor: const Color(0xFFFFFFFF),
+                    shape: RoundedRectangleBorder(
+                      borderRadius: BorderRadius.circular(6),
+                      side: BorderSide(color: Color(0xFF726E6E), width: 1),
+                    ),
+                    padding:
+                        const EdgeInsets.symmetric(horizontal: 12, vertical: 4),
+                    minimumSize: Size(
+                      screenSize.width * 0.67,
+                      screenSize.height * 0.044,
+                    ),
+                  ),
+                  child: const Text(
+                    '취소하기',
+                    style: TextStyle(
+                      fontSize: 13,
+                      fontWeight: FontWeight.w400,
+                      color: Color(0xFF726E6E),
+                    ),
+                  ),
                 ),
               ],
             ),
-          );
-        }
-        return Container();
+          ],
+        );
       },
     );
   }
